Time: 105 ms (34.11%), Space: 76.3 MB (65.70%) - LeetHub

+++ 1514-path-with-maximum-probability/1514-path-with-maximum-probability.java
@@ -14,7 +14,7 @@ class Solution {
         pq.add(new double[]{start,1});
         double[] d=new double[n];
         Arrays.fill(d,0);
-        d[start]=(double)1;
+        d[start]=1;
         while(!pq.isEmpty()){
             double[] node=pq.poll();
             for(double[] arr:G[(int)node[0]]){
